backend/requirements.txt
# FastAPI e servidor ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Banco de dados
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Validação e serialização
pydantic==2.5.3
pydantic-settings==2.1.0

# Autenticação JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilitários
python-dotenv==1.0.0
